requirements.txt
enum34
python-docx
lxml
pymorphy2
nltk